Make Base Class for controller test
